Some test scripts for teleoperation and franka robot interface
