+++ reactive_diffusion_policy/common/space_utils.py
@@ -130,34 +130,3 @@ def homo_matrix_to_pose_9d_batch(mat: np.ndarray) -> np.ndarray:
     pose[:, :3] = mat[:, :3, 3]
     pose[:, 3:9] = mat[:, :3, :2].swapaxes(1, 2).reshape(mat.shape[0], -1)
     return pose
-
-# def pos_rot_to_mat(pos, rot):
-#     shape = pos.shape[:-1]
-#     mat = np.zeros(shape + (4,4), dtype=pos.dtype)
-#     mat[...,:3,3] = pos
-#     mat[...,:3,:3] = rot.as_matrix()
-#     mat[...,3,3] = 1
-#     return mat
-
-# def mat_to_pos_rot(mat):
-#     pos = (mat[...,:3,3].T / mat[...,3,3].T).T
-#     rot = st.Rotation.from_matrix(mat[...,:3,:3])
-#     return pos, rot
-
-# def pos_rot_to_pose(pos, rot):
-#     shape = pos.shape[:-1]
-#     pose = np.zeros(shape+(6,), dtype=pos.dtype)
-#     pose[...,:3] = pos
-#     pose[...,3:] = rot.as_rotvec()
-#     return pose
-
-# def pose_to_pos_rot(pose):
-#     pos = pose[...,:3]
-#     rot = st.Rotation.from_rotvec(pose[...,3:])
-#     return pos, rot
-
-# def pose_to_mat(pose):
-#     return pos_rot_to_mat(*pose_to_pos_rot(pose))
-
-# def mat_to_pose(mat):
-#     return pos_rot_to_pose(*mat_to_pos_rot(mat))
